+++ crates/gen/src/types/class.rs
@@ -1,5 +1,5 @@
 use crate::format_ident;
-use crate::tables::*;
+use crate::winmd::*;
 use crate::types::*;
 use crate::TypeReader;
 use squote::{quote, Literal, TokenStream};
