+++ crates/gen/src/types/delegate.rs
@@ -1,5 +1,5 @@
 use crate::format_ident;
-use crate::tables::*;
+use crate::winmd::*;
 use crate::types::debug;
 use crate::types::*;
 use crate::TypeReader;
