+++ crates/gen/src/types/required_interface.rs
@@ -1,4 +1,4 @@
-use crate::tables::*;
+use crate::winmd::*;
 use crate::types::*;
 use crate::TypeReader;
 use squote::{quote, TokenStream};
