+++ crates/gen/src/winmd/blob.rs
