+++ crates/gen/src/winmd/element_type.rs
@@ -1,4 +1,4 @@
-use crate::blob::Blob;
+use crate::Blob;
 use crate::codes::{Decode, TypeDefOrRef};
 
 #[derive(Debug)]
