+++ crates/gen/src/winmd/interface_impl.rs
